requirements.txt
streamlit==1.38.0
matplotlib==3.9.2
numpy==2.1.3
biopython==1.85
tqdm==4.66.4
orfipy==0.10.0
parasail==1.3.5
